feat(projects): implement comprehensive file management with ZIP upload and custom UI
- Add ZIP file upload functionality with yauzl processing
- Implement file tree component with folder/file structure
- Add custom modal and alert dialogs to replace browser alerts
- Implement toast notifications using sonner
- Add folder deletion functionality
- Fix TypeScript lint warnings and improve type safety
- Update file size display and validation

+++ backend/src/modules/projects/file.entity.ts
@@ -30,7 +30,7 @@ export class File {
     @Column({ length: 50, nullable: true })
     language: string;
 
-    @Column({ type: 'bigint', default: 0 })
+    @Column({ type: 'integer', default: 0 })
     size: number;
 
     @CreateDateColumn({ type: 'timestamptz' })
